+++ api/config/prisma.js
@@ -1,0 +1,3 @@
+const { PrismaClient } = require("@prisma/client");
+
+module.exports = new PrismaClient();
